% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chmmMCMC.R
\docType{class}
\name{Chain-class}
\alias{Chain-class}
\alias{Chain}
\title{A Reference Class to represent a single chain in HMM or CHMM}
\description{
This class corresponds to chain of the HMM.
It performs all the MCMC sampling procedures for the sigle chain.
}
\section{Fields}{

\describe{
\item{\code{observations}}{N x L matrix where rows corresponds to
number of observations while columns corresponds to sequence lengths
MUST be provided by user}

\item{\code{params}}{Parameter list that defines the model parameters.
MUST be supplied by user.
Details of that field are in the package description.}

\item{\code{latent_sequences}}{sampled_latent sequence. It is necessary for model to continue.
It is not necessary for any further action.}

\item{\code{log_likelihoods}}{Log likelihoods of each sequence at each MCMC iteration.
Rows corresponds to different sequences and columns corresponds to iterations}

\item{\code{pi0s}}{Sampled pi0s. Rows corresponds to number of states and columns corresponds iteration number.}

\item{\code{betas}}{Sampled beta parameters. Rows corresponds to beta values and columns corresponds iteration number.}

\item{\code{transitions}}{A list that model needs. Those transitions are converted from corresponding beta iteration.
It is not necessary for any further action.}

\item{\code{emissions}}{Sampled emissions. It is 3 dimensional array and dim[1] corresponds to n_states,
dim[2] corresponds to n_observations and dim[3] corresponds to number of MCMC samples.}

\item{\code{covariates}}{The latent sequences of the other chains. MUST be provided in the updating the chain.}

\item{\code{forward_alpha}}{Forward table. It will be not necessary for any further action.}

\item{\code{diagnosis}}{List that includes MCMC diagnosis. It is the result of the diagnostics() function.}

\item{\code{posterior}}{List that includes posterior probabilities of pi0, transition and emission.
Load observations}
}}


