% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chmmMCMC.R
\docType{class}
\name{Chmm_Mcmc-class}
\alias{Chmm_Mcmc-class}
\alias{Chmm_Mcmc}
\title{A Reference Class to CHMM}
\description{
This class corresponds to CHMM training.
It basically manage the Chain class.
}
\section{Fields}{

\describe{
\item{\code{params}}{Parameter list that defines the model parameters.
MUST be supplied by user.
Details of that field are in the package description.}

\item{\code{chain_names}}{Optional parameter to give the chains a name.}

\item{\code{chains}}{A list that store Chain class objects. It has the size of  number of chains.}

\item{\code{log_likelihoods}}{Log likelihoods of each sequence at each MCMC iteration.
It is the sum of log_likelihoods of all chains.
Rows corresponds to different sequences and columns corresponds to iterations}

\item{\code{n_chains}}{Number of chains. Class calculate it from length of observation list.}

\item{\code{runtime}}{Total time spend on training}

\item{\code{diagnosis}}{Model diagnostics and model scores.}
}}


